Removed broken functionality
Made changes to licensing to reflect the fact that STEC inc. no longer exists, and that this is now a community project.

Removed unimplemented functionality from "notify".
Some of the functionality is available in the EnhanceIO driver but not implemented here. I'd do it, but I have no idea how to convert the C ioctl numbers for Python. Could we consider a C rewrite of eio_cli?